Improve database interactions
I renamed `insert_sample` to `add_sample` since this is a more
descriptive name for the action committed. Similarly, I renamed
`remove_permissions` to `delete_permissions` since `delete` more closely
resembles the words of the rest of the codebase.

Finally, I made `Permissions.from_str` be able to accept muitiple
strings as *args because this closer resembles the idea that a
`Permissions` object is a flag that stores multiple permissions.